mca: add support for component aliasing
This commit adds support for aliasing the name of a component. When a
component name is aliased it can be selected via either the original
or aliased name. Example: if mca_btl_vader_component sets the
mca_component_name_alias to "sm" it will be selected when either -mca
btl sm or -mca btl vader are specified.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>