mca: add support for required components
This commit adds a new flag to the mca_base_component_t structure that
specifies component flags. The only flag currently support is
MCA_BASE_COMPONENT_FLAG_ALLWAYS_CONSIDER which forces the selection
logic in the MCA base to always consider a component whether or not it
was specified. In order to use this support a component must be built
statically and not dynamically. For example, if btl:self is built
statically and sets this flag in its component then -mca btl
vader,openib would provide vader, openib, and self.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>