Make the PMI barrier/modex be non-blocking so that opal_progress can run. This fixes a hang when direct launching jobs that use PSM
Add missing CPPFLAGS from PMI components - thanks to Matias Cabral for pointing it out!

Handle the case where someone specifies --with-pmi and gives us the complete path to the pmi.h file. Abort if we don't find pmi.h or pmi2.h - we were silently continuing on even though the headers were not found!

Thanks to Matias Cabral for his patience in reporting it