opal/cmd_line: enable complete command line parsing
This commit changes the command line parsing behavior when ignoring
both unknown switches and tokens. In this case we are probably trying
to parse part of the command line (-mca, -am, -tune) before the
rest. Before this change the command line parser stopped on the first
unknown switch or token. Now it will continue to parse out
options. This includes options to the app being run by mpirun (unless
-- is specified) but since only -mca, -tune, and -am are affected this
probably is not a big deal. This change only affects the parsing of
command-line MCA parameters in orte-submit, all other parsers are
unchanged.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>