mpool/udreg: manual patch to use mntent
This is a hand-coded patch to udreg mpool
component to use the mntent interface to
parse /proc/mounts entries.

Needed for XC/KNL-xeons running CLE 6
or higher because, depending on use of images
at customer site, there may be some very large
entries in /proc/mounts which broke the original
parsing code, leading to failure to launch, segfaults,
etc.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>