mca/base: add support for filtering components
This commit adds support to the MCA base to filter components based
on a filter this. In this version there are two types of filters:
enable, and disable. If a component's filter matches a disable list
then the component is disabled. If a component matches an enable list
it is used. We will need to further discuss this behavior before it
is ready for master.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>