osc/pt2pt: do not use frag send to send lock request
This commit cleans up some code in the passive target path. The code
used the buffered frag control send path but it is more appropriate to
use the unbuffered one. This avoids checking structures that are
should not be in use in this path.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>

(cherry picked from commit open-mpi/ompi@cb1cb5ffed25f03fabfc4af742e6e066be832a50)

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>